requirements.txt
coverage==7.0.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.23.5
oauthlib==3.2.2
openmeteo_requests==1.1.0
openmeteo_sdk==1.7.1
pandas==1.5.2
Pillow==9.5.0
plotly==5.11.0
pytest==7.2.0
pytest-asyncio==0.20.3
request